+++ src/main/java/com/jcloisterzone/ui/grid/MainPanel.java
@@ -109,7 +109,7 @@ public class MainPanel extends JPanel {
         //TODO add always
         if (capabs.contains(LittleBuildingsCapability.class) ||
             capabs.contains(TunnelCapability.class) ) {
-            gridPanel.addLayer(new TokenLayer(gridPanel, gc));
+            //gridPanel.addLayer(new TokenLayer(gridPanel, gc));
         }
 
         if (capabs.contains(BridgeCapability.class)) {